fix(node-ui): correct useRenewalChaining import paths after the page-layer move
The previous commit landed the git-mv rename but not the import-depth fixes: the
moved hook still referenced `../api.js`/`../stores/tabs.js` (now `../../`) and
LifecycleSection still imported the old `../../pca/useRenewalChaining.js` path.
Repoint both. Verified: full PCA suite green + build:ui exit 0.
